requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
gunicorn
psycopg2-binary
